# src/main/contract/api-specs.yaml
openapi: 3.0.0
info:
  title: Advanced Library REST API
  description: Example about REST in Spring Reactive
  version: 1.0.1
servers:
  - url: http://localhost:8080
    description: Optional server description, e.g. Internal staging server for testing
tags:
  - name: publisher-management
    description: Operations related to publishers.
  - name: author-management
    description: Operations related to authors.
  - name: book-management
    description: Operations related to books.
paths:
  # publishers --------------------------------------------
  /publisher:
    post:
      summary: Register a publisher info.
      description: Register a publisher info.
      tags:
        - publisher-management
      operationId: createPublisher
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublisherInput"
      responses:
        '201':    # status code
          description: The publisher info gets registered.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublisherOutput"
          headers:
            Location:
              description: The URI of the registered publisher info.
              schema:
                type: string
    get:
      summary: Get an array of registered publishers.
      description: Get an array of registered publishers.
      tags:
        - publisher-management
      operationId: getPublishers
      responses:
        '200':    # status code
          description: Returns a list of registered publishers.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PublisherOutput"
  /publisher/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
        description: Publisher id
    get:
      summary: Get publisher info by ID.
      description: Get publisher info by ID.
      tags:
        - publisher-management
      operationId: getPublisherById
      responses:
        '200':    # status code
          description: A JSON representation of the publisher info.
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/PublisherOutput'
  /publisher/{id}/books:
    parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
        description: Publisher id
    get:
      summary: Get books of a given publisher.
      description: Get books of a given publisher.
      tags:
        - publisher-management
      operationId: getPublisherBooks
      responses:
        '200':    # status code
          description: A JSON representation of the publisher info.
          content:
            application/json:
              schema: 
                type: array
                items:
                  $ref: '#/components/schemas/BookOutput'
  # Authors    --------------------------------------------
  /author:
    post:
      summary: Register an author info.
      description: Register an author info.
      tags:
        - author-management
      operationId: createAuthor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthorInput"
      responses:
        '201':    # status code
          description: The author info gets registered.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthorOutput"
          headers:
            Location:
              description: The URI of the registered author info.
              schema:
                type: string
    get:
      summary: Get an array of registered authors.
      description: Get an array of registered authors.
      tags:
        - author-management
      operationId: getAuthors
      responses:
        '200':    # status code
          description: Returns a list of registered authors.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuthorOutput"
  /author/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
        description: Author id
    get:
      summary: Get author info by ID.
      description: Get author info by ID.
      tags:
        - author-management
      operationId: getAuthorById
      responses:
        '200':    # status code
          description: A JSON representation of the author info.
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/AuthorOutput'
  /author/{id}/books:
    parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
        description: Author id
    get:
      summary: Get books of a given author.
      description: Get books of a given author.
      tags:
        - author-management
      operationId: getAuthorBooks
      responses:
        '200':    # status code
          description: A JSON representation of the author s book.
          content:
            application/json:
              schema: 
                type: array
                items:
                  $ref: '#/components/schemas/BookOutput'
  # Books      --------------------------------------------
  /books:
    post:
      summary: Register a new book in the library.
      description: Register a new book in the library.
      tags:
        - book-management
      operationId: create
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookInput"
      responses:
        '201':    # status code
          description: The book gets registered in the library.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookOutput"
          headers:
            Location:
              description: The URI of the registered book
              schema:
                type: string
    get:
      summary: Get an array of book that match the criteria.
      description: Get an array of book that match the criteria.
      tags:
        - book-management
      operationId: getBooksByCriteria
      parameters:
        - in: query
          name: queryParams
          description: Query parameters
          schema:
            $ref: "#/components/schemas/BookQueryParams"
          explode: true
      responses:
        '200':    # status code
          description: Returns a list of books that met the given criteria.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookOutput"
  /books/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
        description: Book id
    get:
      summary: Get book by ID.
      description: Get book by ID.
      tags:
        - book-management
      operationId: getBookById
      responses:
        '200':    # status code
          description: A JSON representation of a book.
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/BookOutput'
    put:
      summary: Update an existing book.
      description: Update an existing book.
      tags:
        - book-management
      operationId: updateOrCreateBook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookInput"
      responses:
        '201':    # status code
          description: The book gets registered in the library.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookOutput"
          headers:
            Location:
              description: The URI of the registered book resource
              schema:
                type: string
        '204':    # status code
          description: Book info gets updated.
    delete:
      summary: Delete an existing book.
      description: Delete an existing book.
      tags:
        - book-management
      operationId: deleteBook
      responses:
        '204':    # status code
          description: The book gets deleted from the repository.
        '404':
          description: The book with given ID does not exist.
  /books/{id}/autor:
    parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
        description: Book id
    get:
      summary: Get author info.
      description: Get author info.
      tags:
        - book-management
      operationId: getBookAuthor
      responses:
        '200':    # status code
          description: A JSON representation of a book.
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/AuthorOutput'
  /books/{id}/publisher:
    parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
        description: Book id
    get:
      summary: Get publisher info.
      description: Get publisher info.
      tags:
        - book-management
      operationId: getBookPublisher
      responses:
        '200':    # status code
          description: A JSON representation of a publisher.
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/PublisherOutput'
components:
  schemas:
    PublisherBase:
      type: object
      properties:
        name:
          type: string
        active:
          type: boolean
        website:
          type: string
          format: uri
    AuthorBase:
      type: object
      properties:
        name:
          type: string
        country:
          type: string
        email:
          type: string
          format: email
    BookBase:
      type: object
      properties:
        title:
          type: string
        authorId:
          type: string
        publisherId:
          type: string
        yearPublished:
          type: integer
    PublisherInput:
      allOf: 
        - $ref: '#/components/schemas/PublisherBase'
      type: object
      required:
        - name
        - active
        - website
    AuthorInput:
      allOf: 
        - $ref: '#/components/schemas/AuthorBase'
      type: object
      required:
        - name
        - country
        - email
    BookInput:
      allOf: 
        - $ref: '#/components/schemas/BookBase'
      type: object
      required:
        - title
        - authorId
        - publisherId
    PublisherOutput:
      allOf: 
        - $ref: '#/components/schemas/PublisherBase'
      type: object
      properties:
        id:
          type: string
      required:
        - id
        - name
        - active
        - website
    AuthorOutput:
      allOf: 
        - $ref: '#/components/schemas/AuthorBase'
      type: object
      properties:
        id:
          type: string
      required:
        - id
        - name
        - country
        - email
    BookOutput:
      allOf: 
        - $ref: '#/components/schemas/BookBase'
      type: object
      properties:
        id:
          type: string
      required:
        - id
        - title
        - authorId
        - publisherId
    BookQueryParams:
      type: object
      properties:
        title:
          type: string
        author:
          type: string
        publisher:
          type: string
        yearPublished:
          type: integer